Changed the way inputs are handled in the network. Now a generator can return a list of inputs and can return None as well (to ignore labels). Those items inside the list are converted in torch tensors one by one. So you can make a generator that yields [images, flow], depth, and both images and flow (which are numpy arrays) are converted in [trImages, trFlow], trDepth. The forward method of the network can just write images, flow = x in forward(self, x) method.

+++ examples/variational-autoencoder/main_vae_classic.py
@@ -21,8 +21,8 @@ class BinaryMNISTReader(MNISTReader):
 		for items in super().iterate_once(type, miniBatchSize):
 			images, _ = items
 			# Images are N(0, I), so we can threshold at 0 to get binary values.
-			images = images > 0
-			yield np.float32(images), np.float32(images)
+			images = np.float32(images > 0)
+			yield images, images
 
 class FCEncoder(NeuralNetworkPyTorch):
 	def __init__(self, numEncodings):
